// src/types/modal.ts
// =================================== MODAL ================================== //

export enum MODAL_NAME {
  NONE = 'NONE',
  PACKAGE_DETAIL = 'PACKAGE_DETAIL',
  TRAVELX = 'TRAVELX',
  ASK_QUESTION = 'ASK_QUESTION',
  HMP_MAP = 'HMP_MAP',
  ASK_A_QUESTION_THANK_YOU ='ASK_A_QUESTION_THANK_YOU',
  AMENITIES_SELECT_MODAL='AMENITIES_SELECT_MODAL'
}